CREW_AI/requirements.txt
crewai==0.28.8
crewai_tools==0.1.6
langchain_community==0.0.29
langchain-groq==0.0.1
python-dotenv==1.0.0
beautifulsoup4>=4.12.3
requests>=2.31.0
streamlit>=1.32.0
langchain>=0.1.0